+++ scripts/ddl/dim_vendors.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS dim_vendors (
-    vendorid INT ENCODE AZ64,
+    vendorid BIGINT ENCODE AZ64,
     vendor_name VARCHAR(100) ENCODE ZSTD,
     updated_by VARCHAR(50) ENCODE ZSTD,
     approved_by VARCHAR(50) ENCODE ZSTD,
